Moved Filter.g and ExpressionTree.  As part of this I had to move PARTITION_DATE_FORMAT from HiveMetaStore to MetaStoreUtils.  Without this the move would have had to wait until I moved HiveMetaStore.  But this caused a cascade of other things to have to wait, so that I would have ended up with one huge patch that moves everything.
