+++ metastore/src/java/org/apache/hadoop/hive/metastore/HiveMetaStore.java
@@ -187,17 +187,6 @@ public class HiveMetaStore extends ThriftHiveMetastore {
   @VisibleForTesting
   static long TEST_TIMEOUT_VALUE = -1;
 
-  /** A fixed date format to be used for hive partition column values. */
-  public static final ThreadLocal<DateFormat> PARTITION_DATE_FORMAT =
-       new ThreadLocal<DateFormat>() {
-    @Override
-    protected DateFormat initialValue() {
-      DateFormat val = new SimpleDateFormat("yyyy-MM-dd");
-      val.setLenient(false); // Without this, 2020-20-20 becomes 2021-08-20.
-      return val;
-    };
-  };
-
   /**
    * default port on which to start the Hive server
    */
