+++ standalone-metastore/src/main/java/org/apache/hadoop/hive/metastore/parser/package-info.java
